+++ src/components/Sidebar.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { LayoutDashboard, Globe, Settings, TrendingUp, Anchor, BarChart2, LogOut, ChevronDown , Users} from "lucide-react";
+import { LayoutDashboard, Globe, Settings, TrendingUp, Anchor, BarChart2, LogOut, ChevronDown, Users } from "lucide-react";
 import { useLanguage } from "@/lib/i18n/LanguageProvider";
 import { usePathname, useRouter } from "next/navigation";
 import { useSession, signOut } from "next-auth/react";